+++ spec/lib/karafka/routing/contracts/topic_spec.rb
@@ -9,7 +9,7 @@ RSpec.describe_current do
       name: "name",
       active: true,
       consumer: Class.new,
-      deserializers: {},
+      deserializing: {},
       kafka: { "bootstrap.servers": "localhost:9092" },
       max_messages: 10,
       max_wait_time: 10_000,
@@ -200,9 +200,9 @@ RSpec.describe_current do
     end
   end
 
-  context "when we validate deserializers" do
+  context "when we validate deserializing" do
     context "when it is not present" do
-      before { config[:deserializers] = nil }
+      before { config[:deserializing] = nil }
 
       it { expect(check).not_to be_success }
     end