+++ spec/support/factories/messages/metadata.rb
@@ -9,11 +9,12 @@ FactoryBot.define do
     partition { 0 }
     timestamp { Time.now.utc }
     deserializers do
-      Karafka::Routing::Features::Deserializers::Config.new(
+      Karafka::Routing::Features::Deserializing::Config.new(
         active: true,
         payload: ->(message) { JSON.parse(message.raw_payload) },
-        key: Karafka::Deserializers::Key.new,
-        headers: Karafka::Deserializers::Headers.new
+        key: Karafka::Deserializing::Deserializers::Key.new,
+        headers: Karafka::Deserializing::Deserializers::Headers.new,
+        parallel: false
       )
     end